+++ ipv4udpclient.c
@@ -10,7 +10,7 @@
 
 
 
-int ipv4udpclient(int port,char*address,int buff_size) 
+int ipv4udpclient(int port,char*address,long unsigned int buff_size) 
 {
     char buf_tx[buff_size];
     int sockfd; 
@@ -33,7 +33,7 @@ int ipv4udpclient(int port,char*address,int buff_size)
     /* assign IP, PORT */
     servaddr.sin_family = AF_INET; 
     servaddr.sin_addr.s_addr = inet_addr(address); 
-    servaddr.sin_port = htons(port); 
+    servaddr.sin_port = htons((uint16_t)port); 
 
     while (1)
     {
