+++ ipv4udpclient.h
@@ -1,1 +1,1 @@
-int ipv4udpclient(int port,char*address,int buff_size);
+int ipv4udpclient(int port,char*address,long unsigned int buff_size);
